Refactor: Decompose optim/__init__.py into focused modules
Split the 1843-line optim/__init__.py into focused, single-responsibility modules:

- geometry.py (183 lines): Vector2D and Polygon classes for 2D calculations
- models.py (233 lines): Data classes (Pin, Component, Spring, Keepout, PlacementConfig)
- placement.py (994 lines): PlacementOptimizer force-directed algorithm
- routing.py (419 lines): FigureOfMerit metrics and RoutingOptimizer

The __init__.py is now 89 lines with only public API re-exports.
All 129 existing tests pass.

Closes #249
